+++ devito/passes/iet/specializations/__init__.py
@@ -1,2 +1,3 @@
 from .C import *  # noqa
 from .openmp import *  # noqa
+from .openacc import *  # noqa
